calB

+++ src/main/java/me/thaithien/associationrules/weighteditem/cai1998mining/db/SimpleDb.java
@@ -110,7 +110,7 @@ public class SimpleDb {
         });
     }
 
-    public long getnTransaction() {
+    public int getnTransaction() {
         return nTrans;
     }
